chore: removed boilerplate class RLBenchEnvAdapter
